Rename workoutStore
