+++ app/(app)/(create-plan)/edit-set.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { View, StyleSheet, TextInput, Button } from "react-native";
 import { ThemedText } from "@/components/ThemedText";
 import { useLocalSearchParams, router } from "expo-router";
-import { useWorkoutStore } from "@/store/store";
+import { useWorkoutStore } from "@/store/workoutStore";
 import { Colors } from "@/constants/Colors";
 
 export default function EditSetScreen() {
