+++ hooks/useCreatePlan.ts
@@ -1,4 +1,4 @@
-import { useWorkoutStore } from "@/store/store";
+import { useWorkoutStore } from "@/store/workoutStore";
 import { insertWorkoutPlan, updateWorkoutPlan } from "@/utils/database";
 import { useEffect, useState } from "react";
 import { Alert } from "react-native";
